Add packhdr program to remove excess data from .exe header
When PKLite compresses a file with the -e flag, it removes this data.  In order
to have a valid uncompressed file to compare against, this program will remove
the extra data from an .exe file and add a PKLite signature, producing an .exe
that should exactly match the decompressed file.